a few code re-org in preparation for adding theme foundations
(cherry picked from commit e2e82eed1292992b64b42888c00c84aaf861249b)
