+++ app/src/main/java/com/jonecx/qio/MainActivity.kt
@@ -2,28 +2,17 @@ package com.jonecx.qio
 
 import android.annotation.SuppressLint
 import android.os.Bundle
-import android.view.ViewGroup
-import android.webkit.WebResourceRequest
-import android.webkit.WebView
-import android.webkit.WebViewClient
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.activity.viewModels
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.viewinterop.AndroidView
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
@@ -32,9 +21,10 @@ import com.jonecx.qio.feature.appstate.AppSessionViewModel
 import com.jonecx.qio.feature.appstate.SessionState
 import com.jonecx.qio.feature.appstate.SessionState.SessionStateLoading
 import com.jonecx.qio.feature.appstate.SessionState.SessionStateValue
+import com.jonecx.qio.feature.authentication.AuthenticatedScreen
+import com.jonecx.qio.feature.authentication.AuthorizationScreen
+import com.jonecx.qio.feature.authentication.ErrorScreen
 import com.jonecx.qio.feature.authentication.LoginViewModel
-import com.jonecx.qio.feature.authentication.OauthUtils.Companion.authorizeRequestUrl
-import com.jonecx.qio.feature.authentication.extractAuthorizationCode
 import com.jonecx.qio.network.ApiResult
 import com.jonecx.qio.network.ApiResult.Error
 import com.jonecx.qio.network.ApiResult.Loading
@@ -44,7 +34,6 @@ import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.launch
-import timber.log.Timber
 
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
@@ -107,59 +96,4 @@ class MainActivity : ComponentActivity() {
             }
         }
     }
-}
-
-@Composable
-fun ErrorScreen() {
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Color.Red),
-        contentAlignment = Alignment.Center,
-    ) {
-        Text(":( failure")
-    }
-}
-
-@Composable
-fun AuthenticatedScreen() {
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Color.Black),
-        contentAlignment = Alignment.Center,
-    ) {
-        Text("Authenticated")
-    }
-}
-
-@Composable
-fun AuthorizationScreen(authorizeAccess: (String) -> Unit) {
-    AndroidView(factory = {
-        WebView(it).apply {
-            this.settings.javaScriptEnabled = true
-            this.settings.useWideViewPort = true
-            this.settings.domStorageEnabled = true
-            this.settings.allowContentAccess = true
-            layoutParams = ViewGroup.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.MATCH_PARENT,
-            )
-            webViewClient = object : WebViewClient() {
-                override fun shouldOverrideUrlLoading(webView: WebView, request: WebResourceRequest): Boolean {
-                    val authorizationCode = request.extractAuthorizationCode()
-                    if (authorizationCode.isNotBlank()) {
-                        Timber.d("Authorization code retrieved")
-                        authorizeAccess(authorizationCode)
-                        return true
-                    }
-                    return false
-                }
-            }
-            Timber.d("Start authorization steps")
-            loadUrl(authorizeRequestUrl)
-        }
-    }, update = {
-        it.loadUrl(authorizeRequestUrl)
-    })
 }
